Master

+++ enums/models.py
@@ -1,5 +1,6 @@
 from django.db import models
 
+
 class ScaleProtocol(models.Model):
     name = models.CharField(primary_key=True,  max_length=50)
 
@@ -8,7 +9,7 @@ class ScaleProtocol(models.Model):
 
 
 class VATRate(models.Model):
-    rate = models.FloatField(primary_key=True)
+    rate = models.IntegerField(primary_key=True)
 
     def __str__(self):
         return f"{self.rate}%"
